Move data encryption
On reflection, all the interpretation of the vba file stuff ought to be
in a separate module that can parse it. I should then be able to test
the code, something I've done none of so far!

This is a first pass on this, moving the VBA encryption algorithms to
their own module. I have also started testing them, although more to do