Finish VBA encryption module
Added in all the tests I meant to & refactored a few things that weren't
right. Most important of these was picking up a bug, where we needed to
do a wrapping_add in the encryption / decryption loop.

The project key is meant to be a checksum on the Project CLSID, so we
could add another method to accept that and encode it properly