Fix hash nulls
I had the wrong order of the bits within each byte for the replace nulls
algorithm within the password hashing section.

I was worried about doing this as I was conscious I was writing
self-confirming code. The bug was found from an actual Excel file that
was encoding nulls, so I now have external confirmation that I am using
the correct bit ordering.